Add files via upload
Random collection of sound effects we can use. Most are sourced from freesound.org and pixabay. So far, the one labeled "PUDDING IS FREEE" is just a really stupid WIP music queue i made using shitty midi horns in garageband during the last meeting- thought it could be funny for when the intro level is complete for an intentionally shitty fanfare thing to happen, but we can see if thats the vibe later or do something else lmao